+++ tests/test_pipe.py
@@ -6,8 +6,12 @@ from functools import partial
 from operator import add
 
 ##############################################################################
+# Pytest imports.
+from pytest import raises
+
+##############################################################################
 # Local imports.
-from bagofstuff.pipe import Pipe
+from bagofstuff.pipe import NoArgument, Pipe
 
 
 ##############################################################################
@@ -40,6 +44,26 @@ def test_pipe_with_pipe() -> None:
     child_pipe = Pipe[int, int](partial(add, 1))
     main_pipe = Pipe[int, int](child_pipe, child_pipe) | child_pipe | child_pipe
     assert main_pipe(1) == 5
+
+
+##############################################################################
+def test_pipe_with_no_initial_argument() -> None:
+    """A pipe should work fine without an initial argument."""
+
+    def one() -> int:
+        return 1
+
+    def plus_one(value: int) -> int:
+        return value + 1
+
+    assert (Pipe[NoArgument, int](one, plus_one) | plus_one)() == 3
+
+
+##############################################################################
+def test_empty_pipeline() -> None:
+    """Calling an empty Pipe should result in an error."""
+    with raises(TypeError):
+        _ = Pipe[None, None]()()
 
 
 ### test_pipe.py ends here