+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # BagOfStuff ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- `Pipe` can now start with zero arguments.
+- An empty `Pipe` when called will now raise a `TypeError`.
+
 ## v0.0.1
 
 **Released: 2026-01-31**